requirements.txt
Django==5.0.7
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
dash==3.2.0
plotly==6.2.0
django-plotly-dash==2.5.0
redis==5.0.6
pandas==2.2.3
#duckdb==1.0.0
qdrant-client==1.8.2
celery[redis]==5.4.0  # background jobs for OEE calculations

# SQL Server support
mssql-django==1.4
pyodbc>=4.0.39

# Data processing and analytics
numpy>=1.24.0
scipy>=1.10.0

# Sparkplug B MQTT Integration
paho-mqtt==1.6.1
eclipse-tahu==1.1.4
protobuf>=3.20.0,<4.0.0

# PLC Communication Libraries
# Siemens S7 Protocol
python-snap7==1.3
# Allen-Bradley EtherNet/IP
pycomm3==1.2.14
# Additional PLC protocols
pylogix==1.4.5

# Async and concurrency
asyncio-mqtt==0.13.0
aiofiles==23.2.1
asyncpg==0.29.0

# Monitoring and observability  
prometheus-client==0.19.0
structlog==23.2.0

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing and simulation
factory-boy==3.3.0
faker==20.1.0

# OPC-UA Client Support
asyncua==1.1.0
cryptography==41.0.7

# GraphQL Support
graphene-django==3.1.5
django-graphql-jwt==0.4.0
django-filter==24.1
django-cors-headers==4.3.1

# Kafka Bridge (Optional)
aiokafka==0.10.0
confluent-kafka==2.3.0

# Time Synchronization
ntplib==0.4.0

# Certificate Management
pyOpenSSL==24.0.0
certifi==2024.2.2
